Music: move UI views into a views/ directory
