+++ apps/src/music/views/SoundsPanel.jsx
