+++ apps/src/music/views/controls.module.scss
