+++ apps/src/music/views/music.module.scss
