+++ apps/src/music/views/soundsPanel.module.scss
