+++ apps/src/music/views/toolbox.module.scss
@@ -1,6 +1,6 @@
 @import "color.scss";
 @import "font.scss";
-@import "../music.module.scss";
+@import "./music.module.scss";
 
 .toolboxCategoryContainer {
   margin-top: 3px;
