+++ apps/src/sites/studio/pages/musiclab/index.js
@@ -1,7 +1,7 @@
 import $ from 'jquery';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import MusicLabView from '@cdo/apps/music/MusicView';
+import MusicLabView from '@cdo/apps/music/views/MusicView';
 
 $(document).ready(function() {
   ReactDOM.render(
